Fix shadow ray bug
Before adding a limit to max t parameter the ray equation can take, an object behind the light source was able to create shadows since it was still intersecting with the ray.

Setting the distance of light as max t value solves this problem since naything behind the light would give a bigger t ray parameter.