Fix a bug in normal thus shading calculation
In the beginning of shading calculations object normals were being flipped if they were facing away from the eye. This was to help the reflections and refractions in dielectric materials. But flipping the normal before the reflection/refraction calculations caused a shading problem. Hence moving the normal flipping a few lines down solves the problem.