Change number of wanted channels to default
Giving 0 for the last argument eliminates a bug which occurs when a
non-zero desired channels is given and also an address is given for the
function to modify to put the number of input images number of channels.
Causing the variable holding number of channels to hold the input images
channel count not the outputted image in the returned pointer.