Separate brdf from the rendering equation
In the previous version compute brdf function was actually doing do
whole shading computation(rendering equation) by itself. Since it is not
the responsibility of the material and brdf classes responsibility to do
this separating them is a good idea. It will also allow us to implement
different type of brdfs